requirements.txt
pyautogui
appdirs
keyring
pyyaml
requests
PyGObject
psutil
